requirements.txt
Flask==1.1.2
gunicorn==19.9.0
line-bot-sdk==1.19.0
pandas==1.1.5
openpyxl==3.0.7
